requirements.txt
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.11
django-cors-headers==4.9.0
Pillow==12.2.0
whitenoise==6.6.0
gunicorn==21.2.0